// demos/ts/with-redux/src/containers/app/types/googleBooksTypes.tsx
export interface BooksState {
    isLoading: boolean;
    books: Book[];
}

export interface StoreState {
    books: BooksState;
}

export interface Book {
    id: string;
    volumeInfo: {
        title: string;
        publisher: string;
        pageCount: number;
        imageLinks: {thumbnail: string};
        publishedDate: string;
        language: string;
        description: string;
        authors: string[];
        averageRating: number;
    };
}

export interface GoogleBooksResponse {
    kind: string;
    totalItems: number;
    items: Book[];
}